Refactor "test_catch_exceptions.py" in two separate files
